requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
opencv-python-headless
numpy
pytesseract
gunicorn
dj-database-url